Move ReactionSystem to its own file
