+++ chempy/__init__.py
@@ -7,7 +7,8 @@ from __future__ import absolute_import, division, print_function
 
 from ._url import __url__
 from ._release import __version__
-from .chemistry import Substance, Reaction, Equilibrium, ReactionSystem, Species, balance_stoichiometry, mass_fractions
+from .chemistry import Substance, Reaction, Equilibrium, Species, balance_stoichiometry, mass_fractions
+from .reactionsystem import ReactionSystem
 from .henry import Henry
 from .util.periodic import atomic_number
