+++ chempy/equilibria.py
@@ -15,7 +15,8 @@ import warnings
 
 import numpy as np
 
-from .chemistry import ReactionSystem, equilibrium_quotient, Equilibrium, Species
+from .chemistry import equilibrium_quotient, Equilibrium, Species
+from .reactionsystem import ReactionSystem
 from ._util import get_backend
 from .util.pyutil import deprecated
 from ._eqsys import EqCalcResult, NumSysLin, NumSysLog, NumSysSquare as _NumSysSquare
